requirements.txt
sentence-transformers==3.4.1
huggingface-hub==0.23.2
langchain
flask
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain_community
langchain_openai
langchain_experimental
-e .